sections/tooling/*.js => sections/tooling/*.md
